Redirect hack moved to galleryconfiguration.
