+++ tests/NuGetGallery.WebUITests.P0/AccountManagement/LogonTest.cs
@@ -1,12 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Globalization;
-using System.Linq;
-using System.Reflection;
 using System.Web.UI;
 using Microsoft.VisualStudio.TestTools.WebTesting;
 using NuGetGallery.FunctionalTests.Helpers;
@@ -20,11 +15,6 @@ namespace NuGetGallery.FunctionalTests.WebUITests.AccountManagement
     public class LogonTest
         : WebTest
     {
-        static LogonTest()
-        {
-            RedirectAssembly("Newtonsoft.Json");
-        }
-
         public LogonTest()
         {
             PreAuthenticate = true;
@@ -47,37 +37,5 @@ namespace NuGetGallery.FunctionalTests.WebUITests.AccountManagement
 
             yield return logonPostRequest;
         }
-
-        /// <summary>
-        /// Source: https://stackoverflow.com/a/32698357
-        /// </summary>
-        public static void RedirectAssembly(string shortName)
-        {
-            ResolveEventHandler handler = null;
-
-            handler = (sender, args) =>
-            {
-                var requestedAssembly = new AssemblyName(args.Name);
-                if (requestedAssembly.Name != shortName)
-                {
-                    return null;
-                }
-
-                var current = AppDomain
-                    .CurrentDomain
-                    .GetAssemblies()
-                    .LastOrDefault(x => x.GetName().Name == shortName);
-
-                if (current != null)
-                {
-                    return current;
-                }
-
-                return null;
-            };
-
-            AppDomain.CurrentDomain.AssemblyResolve += handler;
-        }
     }
 }
-
